Add consensus store
