+++ subscriber/Cargo.toml
@@ -14,6 +14,8 @@ serde = { version = "1.0.136", features = ["derive"] }
 serde_bytes = "0.11.5"
 log = "0.4.14"
 futures = "0.3.21"
+rocksdb = "0.18.0"
+store = { git = "https://github.com/mystenlabs/mysten-infra.git", package = "typed-store", rev = "808de09203d147b43d59114b8afd9e51cbcf5778"}
 
 crypto = { path = "../crypto" }
 primary = { path = "../primary" }
